move client-side message parsing functions into client_messages.py
